backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
openai>=1.3.7
python-dotenv>=1.0.0
pydantic>=2.8.0

# Performance optimization dependencies (optional but recommended)
httpx>=0.25.0  # Connection pooling for OpenAI client
slowapi>=0.1.9  # Rate limiting middleware
redis>=5.0.0  # Redis caching (optional, falls back to in-memory cache)
